vcsim: create underlay network for pNICs in container-backed hosts
Connects sim-hosts to bridges as specified in their config.
The bridges to use for a given pNIC are expressed in the advanced
options using the following pattern (example for pNIC 0):
	RUN.underlay.vmnic0=vcsim-mgmt-underlay

This uses an existing bridge or creates a new one as
needed.

If a host has a container backing, all pNICs defined in the host
template are discarded and new pNICs are created, 1 per underlay
name provided to simulator.HostSystem.configureContainerBacking.
This was the only sane way I found to indicate which bridges a
host should be connected to.

The IP assigned to the container is reflected into the various
host.Config structures associated with the vmknic, eg.
VirtualNicManagerInfo

The simulator.HostSystem.getNetConfigInterface method is prvoided to
allow a caller to retrieve all the various network entities associated
with a NetConfig binding, eg. "management", "vmotion".

Remove use of errors.Join to maintain support for older Go versions.

Known issues:
* podman volume ls filters act as OR instead of AND which results in
  all volumes being deleted any time a single host is removed. Issue
  opened and fixed in podman main.