vcsim: container backing respects changes via reconfigure
If "RUN.container" is added or removed on an existing VM, that change
is applied immediately if the VM is currently powered on.

Modifications to the value of the key do not have an effect unless the
continue needs to be recreated for some reason.

Switches to Go templates for formating docker command output

Includes additional error logging detail