vcsim: add interactive debug github action stage
Adds a stage to the github actions pipeline that provides an ssh server
that allows interactive login to the environment. This only triggers on
failure.

The reason for adding this is due to repeated failures to find
functional arguements for the specific docker version present. Quirks
around the format parameter values specifically.

This is done using the tmate action:
https://github.com/mxschmitt/action-tmate

Corrects boilerplate