chore: Dockerfile & steps for generating types
This patch introduces a Dockerfile for generating the
vim types. The image should be cross-arch compatible.

This patch also enables "gen.sh" to automatically update
the VIM version in "vim25/client.go".

This patch also ensures "gen/rbvmomi/vmodl.db" is downloaded
if it does not already exist when generating types.