vcsim: Remove extra devices in OVF Deploy w/ ConfigSpec
The ConfigSpec returned by CreateImportSpec() will include devices from
the OVF but since the simulator impl just reconfigures the VM with the
user provided ConfigSpec, those default devices won't be removed.  In VM
Operator, likely the only user of this API w/ the ConfigSpec param, this
would result in an extra Reconfigure to remove the unexpected devices.
However, I'd like to be able to assert that a just-created VM matches
the expected config.

It is a easier to just skip the EthCards instead of including all the
disk and disk controllers. A later improvement would be to merge the two
ConfigSpecs so the Reconfigure is not needed but neither is a huge deal
for the simulator.