api: WaitForUpdatesEx & DestroyPropertyFilter
This patch introduces Ex variants for many of the property
collector types/methods to take advantage of the remote API
DestroyPropertyFilter. Now it is possible to use a single
property collector for waiting for updates and remove
previously added property filters.

BREAKING: The semantics around the helper functions in the
          property package have changed. Please review any
          code that calls this package to ensure it is
          compatible with the new behaviors.