api: HardwareVersion.IsSupported vs IsValid
This patch modifies the behavior of the HardwareVersion.IsValid
function so that it returns true if the specified version can be
parsed.

There is a new function, HardwareVersion.IsSupported, that acts
more like the old behavior of IsValid. IsSupported only returns
true when the version is known to GoVmomi.

BREAKING: HardwareVersion.IsValid is more relaxed, consider IsSupported
* HardwareVersion.IsValid returns true if the specified value
  matches a VMX version format.
* HardwareVersion.IsSupported behaves how IsValid used to behave.