nfc: Record the SSL thumbprint for the host the (device) URL refers to
With this change we are recording the SSL thumbprint for the host the device
URL refers to. This is done so that clients of the nfc wrapper can perform
their own SSL verification using the recorded thumbprint if necessary. The
thumbprint is stored in the FileItem struct, but the New constructor is not
modified so as not to introduce any breaking changes.

Please note that the thumbprint is recorded only for the import case, since
that is the only necessary path for our internal usages.