fix: restore govc import.spec DiskProvisioning default value
PR #3427 changed to using generated enum lists in various govc commands.
This also changed the default DiskProvisioning value from "flat" to "monolithicSparse",
which results in a DeviceUnsupportedForVmPlatform fault when importing.

This change restores the default to "flat".
The default values for IPProtocol and IPAllocationPolicy had not changed,
but this commit explicitly sets their values rather than using first element of the generated list.

Fixes #3483