api: ToString for vim types
This patch adds a ToString helper function for invoking on any
vim type. If the specified value is a primitive type, the returned
string is generated using Sprintf. Otherwise the returned string is
created first by attempting to marshal the data using the vimtype
JSON encoder, and if that fails, the stdlib JSON encoder. If that
fails, then fmt.Sprintf("%v") is used.