vcsim: Support VM crypto spec in vC Sim
This patch adds support for "encrypting", "recrypting", and
"decrypting" VMs in vC Sim. Please note, no encryption is
actually used, and this change only impacts the state machine
semantics. For example, if a VM is reconfigured to be encrypted,
then retrieving the VM's "config.keyId" property will return the
ID of the key and provider used to encrypt the VM.