fix: Property Collector updates should be triggered on the empty FilterSet
The PropertyCollector update closure passed to WaitForUpdatesEx is only
triggered when a non-empty FilterSet occurs.
This is an issue when side effects need to be triggered inside the
update function such as notification mechanisms that an update was
acknowledged.

It should be the choice of the update closure to determine whether to
process the empty FilterSet or not.

Closes #3631

Signed-off-by: Brian Rak <brian.rak@broadcom.com>