api: Support encode/decode OVF to JSON
This patch adds support for encoding an `ovf.Envelope` to JSON
using Golang's `encoding/json` package. The resulting output can
be safely decoded back to an `ovf.Envelope`. This decoded envelope
is deep-equal to the one originally decoded from XML.

This feature allows clients to format an OVF envelope as YAML, such
as when part of a Kubernetes API.

Signed-off-by: akutz <akutz@vmware.com>