Add tests for types of meta fault
This adds compound tests for whether a fault is a transient error, or
whether it's an error due to an attempt to power off an already powered
off VM.

The former is widely reimplemented by clients, and often missing some
types. The latter is extremely useful to have when mapping desired
state to imperative.

Signed-off-by: George Hicken <george.hicken@broadcom.com>