Add bindings for cis Task API
Previously, tasks were being unmarshalled into a map and checked for
completion using a const.

This change implements
* all of the status enums,
* the full Info and Progress structs,
* adds utility methods to wait for completion OR wait for RUNNING or
  ERORR

Also fixed a bug where waiting on a task would incur waiting for the
poll interval even if the task was done.

Tested using a simple test file which created a task and waited for
completion or error.

Signed-off-by: Faiyaz Ahmed <faiyaz.ahmed@broadcom.com>