api: Fix OVF defVal & ConfigSpec vAppConfig
This patch addresses some bugs in the way OVFs are unmarshalled as
well as transformed into ConfigSpec objects. For some reason the
original OVF unmarshalling code ignored the fact that properties
have a "value" attribute and not just sub-elements. In fact, the latter
is not even in the DMTF OVF spec. There was also a bug in how the
default value and value were treated as the same, when they are not.

Signed-off-by: akutz <andrew.kutz@broadcom.com>