api: Update ofvEnvelop.ToConfigSpec and configInfo.ToConfigSpec to handle nil ovfProperty userConfigurable
When creating a VM we want to be sure that VApp properties are being
configured correctly. We have seen behavior where a nil UserConfigurable
field is treated as true. To mitigate this behavior, this PR updates the
ToConfigSpec functions in configspec.go and helpers.go to explicitly set
the value of a Property's UserConfigurable field to false if the field is
nil. This will ensure that the property is correctly set to UserConfigurable: false
when creating a VM.

Signed-off-by: Adnan Baruni <adnan.baruni@broadcom.com>