vcsim: Consider default devs from VM spec
This patch updates the way vC Sim creates a new VM. Previously the
VM's default devices were always added, regardless of whether or not
the incoming ConfigSpec specified any of the default device info.

This patch instructs vC Sim to consider the incoming ConfigSpec and
check if it contains one or more of the default devices, using them
from the ConfigSpec if it does.

Signed-off-by: akutz <andrew.kutz@broadcom.com>